pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openai_tts_api"
version = "0.1.0"
description = "API for generating audio from openAI text to speech."
authors = [
    { name = "Laurens Veltman"}
]
dependencies = [
    "requests",
    "python-decouple",
    "fastapi"
]
requires-python = ">=3.10"
readme = "README.md"
license = "MIT"